# esp32: Update ESP-IDF to v5.2.2, optimise size