--- a/ports/esp32/boards/sdkconfig.base
+++ b/ports/esp32/boards/sdkconfig.base
@@ -121,3 +121,9 @@ CONFIG_ETH_USE_SPI_ETHERNET=y
 CONFIG_ETH_SPI_ETHERNET_W5500=y
 CONFIG_ETH_SPI_ETHERNET_KSZ8851SNL=y
 CONFIG_ETH_SPI_ETHERNET_DM9051=y
+
+# Using newlib "nano" formatting saves size on SoCs where "nano" formatting
+# functions are in ROM. Note some newer chips (c2,c6) have "full" newlib
+# formatting in ROM instead and should override this, check
+# ESP_ROM_HAS_NEWLIB_NANO_FORMAT.
+CONFIG_NEWLIB_NANO_FORMAT=y