--- a/ports/esp32/mphalport.c
+++ b/ports/esp32/mphalport.c
@@ -100,7 +100,7 @@ void check_esp_err_(esp_err_t code, const char *func, const int line, const char
 
 uintptr_t mp_hal_stdio_poll(uintptr_t poll_flags) {
     uintptr_t ret = 0;
-    #if CONFIG_ESP_CONSOLE_USB_SERIAL_JTAG
+    #if CONFIG_ESP_CONSOLE_USB_SERIAL_JTAG_ENABLED
     usb_serial_jtag_poll_rx();
     #endif
     if ((poll_flags & MP_STREAM_POLL_RD) && stdin_ringbuf.iget != stdin_ringbuf.iput) {
@@ -114,7 +114,7 @@ uintptr_t mp_hal_stdio_poll(uintptr_t poll_flags) {
 
 int mp_hal_stdin_rx_chr(void) {
     for (;;) {
-        #if CONFIG_ESP_CONSOLE_USB_SERIAL_JTAG
+        #if CONFIG_ESP_CONSOLE_USB_SERIAL_JTAG_ENABLED
         usb_serial_jtag_poll_rx();
         #endif
         int c = ringbuf_get(&stdin_ringbuf);
@@ -133,7 +133,7 @@ mp_uint_t mp_hal_stdout_tx_strn(const char *str, size_t len) {
     if (release_gil) {
         MP_THREAD_GIL_EXIT();
     }
-    #if CONFIG_ESP_CONSOLE_USB_SERIAL_JTAG
+    #if CONFIG_ESP_CONSOLE_USB_SERIAL_JTAG_ENABLED
     usb_serial_jtag_tx_strn(str, len);
     did_write = true;
     #elif CONFIG_USB_OTG_SUPPORTED