--- a/tools/ci.sh
+++ b/tools/ci.sh
@@ -116,7 +116,7 @@ function ci_cc3200_build {
 # ports/esp32
 
 # GitHub tag of ESP-IDF to use for CI (note: must be a tag or a branch)
-IDF_VER=v5.0.4
+IDF_VER=v5.2.2
 
 export IDF_CCACHE_ENABLE=1