… ports/esp32/boards/ESP32_GENERIC_C3/sdkconfig.c3usb
… ports/esp32/boards/LOLIN_C3_MINI/sdkconfig.board
… ports/esp32/machine_pin.h
… ports/esp32/uart.h
… ports/esp32/usb.c