### Summary

This PR updates the esp32 port's ESP-IDF version to v5.2.2 and changes the version build in CI from v5.0.4 to v5.2.2.

This is in anticipation of adding support for new SoCs ESP32-C6 (see #11869) and ESP32-C2 (see #15440) in the linked PRs.

Propose after merging this and changing the nightly version then we wait a couple of weeks and if no regressions are reported then remove support for older ESP-IDF, including some code cleanup to suit.

### Code Size

The usual problem with updating ESP-IDF is an increase in code size. This PR applies some config changes to reduce resource usage, and also adds a utility script to quickly measure the static resource usage of multiple boards & ESP-IDF versions.

Current master branch resource usage:

| BOARD            | BOARD_VARIANT | IDF Version | Binary Size | Static IRAM Size | Static DRAM Size |
|------------------|---------------|-------------|-------------|------------------|------------------|
| ESP32_GENERIC    |               | v5.0.4      |     1669037 |           128286 |            52892 |
|                  |               | v5.0.5      |     1682469 |           125822 |            53732 |
|                  |               | v5.1.2      |     1708393 |           129398 |            54276 |
|                  |               | v5.2        |     1707709 |           131150 |            55588 |
|                  |               | v5.2.2      |     1663573 |           121966 |            55236 |
| ESP32_GENERIC    | D2WD          | v5.0.4      |     1425121 |           121270 |            52768 |
|                  |               | v5.0.5      |     1435673 |           118666 |            53608 |
|                  |               | v5.1.2      |     1456773 |           121306 |            54160 |
|                  |               | v5.2        |     1446181 |           121578 |            54776 |
|                  |               | v5.2.2      |     1399309 |           112318 |            54340 |
| ESP32_GENERIC    | SPIRAM        | v5.0.4      |     1545837 |           130618 |            53308 |
|                  |               | v5.0.5      |     1557441 |           127802 |            54124 |
|                  |               | v5.1.2      |     1551761 |           128770 |            54540 |
|                  |               | v5.2        |     1543413 |           126870 |            55208 |
|                  |               | v5.2.2      |     1497321 |           117170 |            54788 |
| ESP32_GENERIC_S3 |               | v5.0.4      |     1564905 |           108674 |            49080 |
|                  |               | v5.0.5      |     1581793 |           104290 |            49344 |
|                  |               | v5.1.2      |     1595049 |            16383 |           138071 |
|                  |               | v5.2        |     1622129 |            16383 |           140159 |
|                  |               | v5.2.2      |     1632013 |            16383 |           143315 |
| ESP32_GENERIC_S3 | SPIRAM_OCT    | v5.0.4      |     1568157 |           110966 |            49980 |
|                  |               | v5.0.5      |     1585081 |           106582 |            50236 |
|                  |               | v5.1.2      |     1598289 |            16383 |           141267 |
|                  |               | v5.2        |     1625693 |            16383 |           143743 |
|                  |               | v5.2.2      |     1635609 |            16383 |           146923 |

Note that the ESP32-S3 pools IRAM and DRAM, so the last two columns need to be added together to be meaningful.

Resource usage after applying changes in this PR:

 BOARD | BOARD_VARIANT | IDF Version | Binary Size | Static IRAM Size | Static DRAM Size |
|-------|---------------|-------------|-------------|------------------|------------------|
| ESP32_GENERIC |  | v5.2.2 | 1627321 | 115310 | 55236 |
| ESP32_GENERIC | D2WD | v5.2.2 | 1361777 | 108074 | 54332 |
| ESP32_GENERIC | SPIRAM | v5.2.2 | 1459885 | 113046 | 54780 |
| ESP32_GENERIC_S3 |  | v5.2.2 | 1574649 | 16383 | 135299 |
| ESP32_GENERIC_S3 | SPIRAM_OCT | v5.2.2 | 1578245 | 16383 | 138907 |

Summary:

* For ESP32. the total binary size and the IRAM usage is now *smaller* on v5.2.2 with these changes compared to master branch with ESP-IDF v5.0.4.
* For ESP32-S3 the total binary size got bigger moving from v5.0.4 to v5.2.2(!) but after applying these changes it's  only a little bigger than ESP-IDF v5.0.4. Not sure why this is...
* About 2344 less bytes of DRAM is available at runtime for ESP-IDF heap, compared to 5.0.4. As far as I can tell this is because of more "iram safe" error log strings that are linked to static DRAM. I can't see any easy way to tackle this, unfortunately.

### Testing

* [x] Run unit tests on ESP32 and ESP32-S3
* [x] Add new unit test for SPI in linked PR
* [x] Run multi_bluetooth and multi_net tests on ESP32 and ESP32-S3
* [x] Test ESP32-C3 with built-in USB device

### Other TODOs

* [x] Resolve mbedTLS 3.6 upgrade fix
* [x] Merge stack check fix for ESP32-C3 https://github.com/micropython/micropython/pull/15605

### Trade-offs and Alternatives

The main trade-off decision is keeping vs dropping support for older ESP-IDF, particularly as there is a small impact to available DRAM on v5.2.2. However supporting three different minor versions (5.0, 5.1, 5.2) is a lot of additional overhead for MicroPython, and the difference between 5.1 and 5.2 doesn't seem like it really justifies keeping 5.1 support around. Let me know if I'm wrong!