@projectgus 2024-07-23
Rebased to pick up the changes in #15476, the numbers in the main post will be slightly wrong now.