@projectgus 2024-07-23
I've started a discussion on Discord about more completely dropping support for older ESP-IDF in this PR, as there are a bunch of compatibility checks that we can simplify out:
https://discord.com/channels/574275045187125269/952412628926472273/1265452441932136550

Comments welcome.