@projectgus 2024-07-25
Have run unit tests, multi_bluetooth and multi_net (Wi-Fi) tests on ESP32 and ESP32-S3. Found one issue introduced by updating from v5.0.4 to v5.2.2 (mbedTLS-related commit in this PR), plus several other broken tests or bugs that were pre-existing (to be submitted in separate PRs).

With all these changes present I now see no test failures, apart from some quite infrequent stress test failures. :tada: 

I wasn't able to test on ESP32-C3, both the boards I have are quite old (probably predating the C3's public release) and don't boot reliably on v5.2.2. I see some debug logs about missing eFuse calibration data, so I assume this is an issue with my boards and not MicroPython. Have ordered a newer C3 board, but if someone has a C3 at hand do they mind checking that it boots OK please?

@dpgeorge Based on @andrewleech's observation in Discord that some people ran into Wi-Fi issues with v5.2, I suggest we keep all versions supported in this PR but switch CI and nightly builds to v5.2.2. If no significant reports are received in a couple of weeks then I'll submit another PR to drop support for older versions and cleanup the code to suit. What do you think?