@dpgeorge 2024-07-25
> if someone has a C3 at hand do they mind checking that it boots OK please

I will check a C3.

> I suggest we keep all versions supported in this PR but switch CI and nightly builds to v5.2.2. If no significant reports are received in a couple of weeks then I'll submit another PR to drop support for older versions and cleanup the code to suit. What do you think?

Sounds good!