@andrewleech 2024-07-25
I've just compiled and flashed this to a new C3 board (recently received, been meaning to try it) [using official IDF v5.2.2 docker](https://github.com/micropython/micropython/wiki/Compiling-esp32-port-with-docker)

```
anl@STEP: ~/micropython $ esp-compile v5.2.2 ESP32_GENERIC_C3 /dev/ttyACM0
... snip ...
Wrote 3072 bytes at 0x00008000 in 0.0 seconds (599.7 kbit/s)...
Hash of data verified.

Leaving...
Hard resetting via RTS pin...
Done
make: Leaving directory '/home/anl/micropython/ports/esp32'

anl@STEP: ~/micropython $ mpremote
Connected to MicroPython at /dev/ttyACM0
Use Ctrl-] or Ctrl-x to exit this shell

>>>
MPY: soft reboot
MicroPython v1.24.0-preview.137.gec63829b07.dirty on 2024-07-25; ESP32C3 module with ESP32C3
Type "help()" for more information.
>>>
```