@andrewleech 2024-07-25
I just ran multi tests between a C3 and S3 (FLASH_4M), both running 5.2.2 @ ec63829b07:
```
anl@STEP: ~/micropython/tests $ ./run-multitests.py -ipyb:/dev/ttyACM0 -ipyb:/dev/ttyACM1 multi_net/*.py
multi_net/asyncio_tcp_client_rst.py on ttyACM0|ttyACM1: skip
multi_net/asyncio_tcp_close_write.py on ttyACM0|ttyACM1: pass
multi_net/asyncio_tcp_readall.py on ttyACM0|ttyACM1: pass
multi_net/asyncio_tcp_readexactly.py on ttyACM0|ttyACM1: pass
multi_net/asyncio_tcp_readinto.py on ttyACM0|ttyACM1: pass
multi_net/asyncio_tcp_server_client.py on ttyACM0|ttyACM1: pass
multi_net/asyncio_tls_server_client.py on ttyACM0|ttyACM1: skip
multi_net/asyncio_tls_server_client_cert_required_error.py on ttyACM0|ttyACM1: skip
multi_net/asyncio_tls_server_client_readline.py on ttyACM0|ttyACM1: skip
multi_net/asyncio_tls_server_client_verify_error.py on ttyACM0|ttyACM1: skip
multi_net/ssl_cert_ec.py on ttyACM0|ttyACM1: skip
multi_net/ssl_cert_rsa.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_check_hostname_error.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_getpeercert.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_server_client.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_server_client_ciphers.py on ttyACM0|ttyACM1: FAIL
### TEST ###
--- instance0 ---

Traceback (most recent call last):
  File "<stdin>", line 16, in <module>
OSError: [Errno 2] ENOENT

--- instance1 ---

### TRUTH ###
--- instance0 ---
b'client to server'
--- instance1 ---
b'server to client'
### DIFF ###
--- /tmp/tmp5czsrt2w    2024-07-25 16:09:40.301553122 +1000
+++ /tmp/tmpgz1a7lb_    2024-07-25 16:09:40.301553122 +1000
@@ -1,4 +1,8 @@
 --- instance0 ---
-b'client to server'
+
+Traceback (most recent call last):
+  File "<stdin>", line 16, in <module>
+OSError: [Errno 2] ENOENT
+
 --- instance1 ---
-b'server to client'
+
multi_net/sslcontext_server_client_files.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_verify_callback.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_verify_error.py on ttyACM0|ttyACM1: skip
multi_net/sslcontext_verify_time_error.py on ttyACM0|ttyACM1: skip
multi_net/tcp_accept_recv.py on ttyACM0|ttyACM1: pass
multi_net/tcp_client_rst.py on ttyACM0|ttyACM1: skip
multi_net/tcp_data.py on ttyACM0|ttyACM1: pass
multi_net/udp_data.py on ttyACM0|ttyACM1: pass
24 tests performed
8 tests passed
15 tests skipped: multi_net/asyncio_tcp_client_rst.py multi_net/asyncio_tls_server_client.py multi_net/asyncio_tls_server_client_cert_required_error.py multi_net/asyncio_tls_server_client_readline.py multi_net/asyncio_tls_server_client_verify_error.py multi_net/ssl_cert_ec.py multi_net/ssl_cert_rsa.py multi_net/sslcontext_check_hostname_error.py multi_net/sslcontext_getpeercert.py multi_net/sslcontext_server_client.py multi_net/sslcontext_server_client_files.py multi_net/sslcontext_verify_callback.py multi_net/sslcontext_verify_error.py multi_net/sslcontext_verify_time_error.py multi_net/tcp_client_rst.py
1 tests failed: multi_net/sslcontext_server_client_ciphers.py
```

Ah I see that failed test requires extra files created/copied onto the devices, my bad