@dpgeorge 2024-07-25
> I'm guessing you need to set the time on the devices somehow

Indeed!  Try `mpremote rtc --set`.