@andrewleech 2024-07-25
> 
> Indeed! Try `mpremote rtc --set`.

Oh that's easier than I was guessing, I forgot about that command thanks!
I've updated my readme pr with an appropriate note.