@dpgeorge 2024-07-26
I tested this PR with the two C3 boards that I have:
- ESP32-C3-DevKitC-02
- ESP32-C3-DevKitM-1

And IDF 5.2.2 with this PR does not work on either of them.

Looking at the DevKitC in particular, which has a ESP32-C3-WROOM-02 module.  With this PR and IDF 5.2.2 the output on the /dev/ttyUSB0 serial when I press the RST button is:
```
$ mpremote
Connected to MicroPython at /dev/ttyUSB0
Use Ctrl-] or Ctrl-x to exit this shell
ESP-ROM:esp32c3-api1-20210207
Build:Feb  7 2021
rst:0x1 (POWERON),boot:0xc (SPI_FAST_FLASH_BOOT)
SPIWP:0xee
mode:DIO, clock div:1
load:0x3fcd5820,len:0xf28
load:0x403cc710,len:0x89c
load:0x403ce710,len:0x2b14
entry 0x403cc710
```

When I use this PR but with IDF 5.0.5 I get:
```
$ mpremote
Connected to MicroPython at /dev/ttyUSB0
Use Ctrl-] or Ctrl-x to exit this shell
ESP-ROM:esp32c3-api1-20210207
Build:Feb  7 2021
rst:0x1 (POWERON),boot:0xc (SPI_FAST_FLASH_BOOT)
SPIWP:0xee
mode:DIO, clock div:1
load:0x3fcd5820,len:0xe8c
load:0x403cc710,len:0x6ec
load:0x403ce710,len:0x2b24
entry 0x403cc710
MicroPython v1.24.0-preview.155.g768bdf99d on 2024-07-26; ESP32C3 module with ESP32C3
Type "help()" for more information.
>>> 
```

These are pretty old boards, but I think they should still work with IDF 5.2.2, considering they work with IDF 5.0.5.  Comparing to @andrewleech 's C3 test above:
- the DevKitC takes about twice as long to download firmware: Wrote 1602560 bytes at 0x00010000 in 56.3 seconds (227.6 kbit/s)...
- it uses /dev/ttyUSB0 (not /dev/ttyACM0)