@andrewleech 2024-07-26
How interesting! With DevKitC is that via a UART bridge?

My board is a cheap ESP32 C3 SuperMini which only exposes the built-in usb, this drops out on rst button and comes up as ACM.